updating to block redirects
